feat: liquidity provider dashboard - display multiple fees
