--- libs/liquidity-provision/README.md
@@ -1,7 +1,0 @@
-# liquidity-provision
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test liquidity-provision` to execute the unit tests via [Jest](https://jestjs.io).
