--- libs/liquidity-provision/project.json
@@ -1,23 +1,0 @@
-{
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/liquidity-provision/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/liquidity-provision/**/*.{ts,tsx,js,jsx}"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/liquidity-provision"],
-      "options": {
-        "jestConfig": "libs/liquidity-provision/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  }
-}
