--- libs/liquidity-provision/src/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './utils';
-export * from './liquidity-provision-provider';
