+++ tsconfig.base.json
@@ -27,9 +27,6 @@
       "@vegaprotocol/governance": ["libs/governance/src/index.ts"],
       "@vegaprotocol/ledger": ["libs/ledger/src/index.ts"],
       "@vegaprotocol/liquidity": ["libs/liquidity/src/index.ts"],
-      "@vegaprotocol/liquidity-provision": [
-        "libs/liquidity-provision/src/index.ts"
-      ],
       "@vegaprotocol/market-depth": ["libs/market-depth/src/index.ts"],
       "@vegaprotocol/market-info": ["libs/market-info/src/index.ts"],
       "@vegaprotocol/market-list": ["libs/market-list/src/index.ts"],
