+++ workspace.json
@@ -16,7 +16,6 @@
     "governance": "libs/governance",
     "ledger": "libs/ledger",
     "liquidity": "libs/liquidity",
-    "liquidity-provision": "libs/liquidity-provision",
     "liquidity-provision-dashboard": "apps/liquidity-provision-dashboard",
     "liquidity-provision-dashboard-e2e": "apps/liquidity-provision-dashboard-e2e",
     "market-depth": "libs/market-depth",
